# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "kash-server",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.2.0"
  },
  "paths": {
...
    "/records": {
      "get": {
        "tags": [
          "records"
        ],
        "operationId": "get_records",
        "parameters": [
          {
            "name": "start_date",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "end_date",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "limit",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "offset",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Records retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetRecordsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": [
          "records"
        ],
        "operationId": "create_record",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRecordPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Record created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Record"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/records/{id}": {
      "put": {
        "tags": [
          "records"
        ],
        "operationId": "update_record",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "record id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRecordPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Record updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Record"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "records"
        ],
        "operationId": "delete_record",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "record id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Record deleted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateRecordPayload": {
        "type": "object",
        "required": [
          "name",
          "amount",
          "currency",
          "category_id",
          "date"
        ],
        "properties": {
          "amount": {
            "type": "number",
            "format": "double"
          },
          "category_id": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "GetRecordsResponse": {
        "type": "object",
        "required": [
          "records",
          "total_count"
        ],
        "properties": {
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Record"
            }
          },
          "total_count": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "Record": {
        "type": "object",
        "required": [
          "id",
          "name",
          "amount",
          "currency",
          "date"
        ],
        "properties": {
          "amount": {
            "type": "number",
            "format": "double"
          },
          "category_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "currency": {
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "UpdateRecordPayload": {
        "type": "object",
        "properties": {
          "amount": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "category_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "date": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "axum_session"
      }
    }
  },
  "security": [
    {
      "session": []
    }
  ]
}